Add upgrade step to migrate from legacy logs to standard
Rough matching from legacy logs to standard log store, based on timestamp.
The newest record in the standard log store which matches the timestamp of
the last exported log ID will be used, so some actions (those which happen
within the same second as the last exported entry) may not be captured.