Add new setting to control resource sort order
Resource sort order changed from course-order (i.e. the order in which the
resources appear within the course) to alphabetic, grouped by type.  This
new option allows the user to choose between the two, though course-order
now has the addition of the resource type toggles at the top of the list.